implement permessions

+++ application/modules/patients/views/manage.php
@@ -14,7 +14,9 @@
 			<button type="button" class="btn btn-primary">Delete</button>
 			<button type="button" class="btn btn-primary">Update</button>
 			-->
-			<button type="button" data-original-title="<?php echo $this->lang->line('__common_create_new');?>" class="create btn btn-primary btn-sm"><i class="fa fa-plus"></i> <?php echo $this->lang->line('__common_create');?></button>
+			<?php if(($this->admin_role_id != $this->role_id) ? $this->Role->has_permission('patients', $this->role_id, 'create',  $this->client_id) : true) { ?>
+				<button type="button" data-original-title="<?php echo $this->lang->line('__common_create_new');?>" class="create btn btn-primary btn-sm"><i class="fa fa-plus"></i> <?php echo $this->lang->line('__common_create');?></button>
+			<?php } ?>
 		</div>
 	</div>
 </div>
@@ -67,6 +69,9 @@ the <section></section> and you can use wells or panels instead
 <!-- end widget grid -->
 
 <script type="text/javascript">
+	var can_view = 	'<?php echo ($this->admin_role_id != $this->role_id) ? $this->Role->has_permission('patients', $this->role_id, 'view',   $this->client_id) : true; ?>';
+	var can_update = '<?php echo ($this->admin_role_id != $this->role_id) ? $this->Role->has_permission('patients', $this->role_id, 'update',   $this->client_id) : true; ?>';
+	var can_delete = '<?php echo ($this->admin_role_id != $this->role_id) ? $this->Role->has_permission('patients', $this->role_id, 'delete',   $this->client_id) : true; ?>';
 
 	$(".create").click(function (e) {
 		var title = $(this).attr('data-original-title');
@@ -326,7 +331,12 @@ the <section></section> and you can use wells or panels instead
 							}else{
 								bday = '--';
 							}
-		                    newData = '<a rel="tooltip" data-placement="top" data-original-title="<?php echo $this->lang->line('__common_details');?>" href="'+user_link+'/'+row['id']+'">'+ row['fullname'] + '</a>';
+							if(can_view){
+								newData += '<a rel="tooltip" data-placement="top" data-original-title="<?php echo $this->lang->line('__common_details');?>" href="'+user_link+'/'+row['id']+'">'+ row['fullname'] + '</a>';
+							}else{
+								newData += row['fullname']
+							}
+		                   
 							newData += '<br>'+ row['birthday'];
 		                    
 		                    return newData;
@@ -415,12 +425,13 @@ the <section></section> and you can use wells or panels instead
 		                // this case `data: 4`.
 		                "render": function (data, type, row) {
 		                    newData = "";
-		                    newData = '<a rel="tooltip" data-placement="left" data-original-title="<?php echo $this->lang->line('__common_reset_password');?>" href="'+BASE_URL+'patients/reset/'+row['id']+'/" class="hidden bootbox btn btn-default btn-xs"><i class="fa fa-key"></i></a>&nbsp;';
-		                   
-							newData += '<a rel="tooltip" data-placement="bottom" data-original-title="<?php echo $this->lang->line('__common_delete');?>" href="'+BASE_URL+'patients/delete/'+row['id']+'/" class="delete"><i class="far fa-trash-alt fa-lg"></i></a>&nbsp;';
-						
-							newData += '<a rel="tooltip" data-placement="bottom" data-original-title="<?php echo $this->lang->line('__common_update');?>"  href="'+BASE_URL+'patients/view/'+row['id']+'/" class="bootbox"><i class="far fa-edit fa-lg"></i></a>&nbsp;';
 		                    
+							if(can_delete){
+								newData += '<a rel="tooltip" data-placement="bottom" data-original-title="<?php echo $this->lang->line('__common_delete');?>" href="'+BASE_URL+'patients/delete/'+row['id']+'/" class="delete"><i class="far fa-trash-alt fa-lg"></i></a>&nbsp;';
+							}
+							if(can_update){
+								newData += '<a rel="tooltip" data-placement="bottom" data-original-title="<?php echo $this->lang->line('__common_update');?>"  href="'+BASE_URL+'patients/view/'+row['id']+'/" class="bootbox"><i class="far fa-edit fa-lg"></i></a>&nbsp;';
+							}
 							newData += '<a rel="tooltip" data-placement="bottom" data-original-title="<?php echo $this->lang->line('__common_records');?>" href="'+BASE_URL+'patients/decoded/medications/'+row['id']+'/" class="move "><i class="fas fa-book fa-lg"></i></a>&nbsp;';
 		                    return newData;
 		                },
